requirements.txt
PySide6>=6.5.0
Pillow>=9.5.0
numpy>=1.24.0
cryptography>=41.0.0
pyinstaller>=5.13.0
pytest>=7.4.0
black>=23.7.0
flake8>=6.0.0
sphinx>=7.1.0
